installer: Updated installer to be more robust and verbose.
Ensure running on Linux for now, fall back onto curl if no wget, suggest installation to /usr/local/bin if possible, warn if overwriting a previous installation, ask to append to ~/.profile if writeable, and provide better next steps notification.

+++ public/install.sh
@@ -3,42 +3,71 @@
 
 # URL to fetch updates from.
 STATEURL="https://s3.ca-central-1.amazonaws.com/cli-update/update/state/"
-# Name of the executable to ultimately
+# Name of the executable to ultimately use.
 STATEEXE="state"
 # ID of the $PATH entry in the user's ~/.profile for the executable.
-STATEID="ActiveState"
+STATEID="ActiveStateCLI"
 
-# Determine the name of the State binary package to install.
-case `uname -m` in
-i?86)
-  statejson=linux-386.json
-  statepkg=linux-386.gz
-  stateexe=linux-386
+# Determine the current OS.
+case `uname -s` in
+Linux)
+  os=linux
   ;;
-x86_64)
-  statejson=linux-amd64.json
-  statepkg=linux-amd64.gz
-  stateexe=linux-amd64
+*BSD)
+  os=`uname -s | tr '[A-Z]' '[a-z]'`
+  echo "BSDs not supported yet"
+  exit 1
+  ;;
+Darwin)
+  os=darwin
+  echo "MacOS not supported yet"
+  exit 1
   ;;
 *)
-  echo "Unknown architecture: `uname -m`"
+  echo "Unsupported OS: `uname -s`"
   exit 1
-;;
+  ;;
 esac
+
+# Determine the current architecture.
+case `uname -m` in
+i?86)
+  arch=386
+  ;;
+x86_64)
+  arch=amd64
+  ;;
+esac
+
+# Construct system-dependent filenames.
+statejson=$os-$arch.json
+statepkg=$os-$arch.gz
+stateexe=$os-$arch
 
 echo "Preparing for installation..."
 
 if [ ! -f $statepkg ]; then
-  # Determine the latest version to fetch.
-  version=`wget -q -O - $STATEURL$statejson | grep -m 1 '"Version":' | awk '{print $2}' | tr -d '",'`
-  echo "Fetching the latest version: $version"
-  # Fetch it.
-  wget -q ${STATEURL}${version}/${statepkg} || exit 1
+  if [ ! -z "`which wget`" ]; then
+    # Determine the latest version to fetch.
+    version=`wget -q -O - $STATEURL$statejson | grep -m 1 '"Version":' | awk '{print $2}' | tr -d '",'`
+    echo "Fetching the latest version: $version"
+    # Fetch it.
+    wget -q ${STATEURL}${version}/${statepkg} || exit 1
+  elif [ ! -z "`which curl`" ]; then
+    # Determine the latest version to fetch.
+    version=`curl -s $STATEURL$statejson | grep -m 1 '"Version":' | awk '{print $2}' | tr -d '",'`
+    echo "Fetching the latest version: $version"
+    # Fetch it.
+    curl -s -o $statepkg ${STATEURL}${version}/${statepkg} || exit 1
+  else
+    echo "Either wget or curl is required to download files"
+    exit 1
+  fi
 fi
 
 # Extract the State binary.
 echo "Extracting $statepkg..."
-gunzip $statepkg
+gunzip $statepkg || exit 1
 
 # Verify checksum.
 echo "Verifying checksum..."
@@ -51,20 +80,37 @@ if [ "`sha256sum -b $stateexe | cut -d ' ' -f1`" != "$shasum" ]; then
   exit 1
 fi
 
-# Prompt the user for a directory to install to.
-installdir="`pwd`"
+# Prompt the user for a directory to install to, defaulting to /usr/local/bin if
+# the user has write permissions, else a local bin.
+if [ -w "/usr/local/bin" ]; then
+  installdir="/usr/local/bin"
+else
+  installdir="$HOME/.local/bin"
+fi
 while "true"; do
   echo -n "Please enter the installation directory [$installdir]: "
   read input
   if [ -e "$input" -a ! -d "$input" ]; then
-    echo "$input exists and is not a directory."
+    echo "$input exists and is not a directory"
+    continue
+  elif [ -e "$input" -a ! -w "$input" ]; then
+    echo "You do not have permission to write to $input"
     continue
   fi
   if [ ! -z "$input" ]; then
-    installdir="`realpath \"$input\"`"
+    if [ ! -z "`realpath \"$input\" 2>/dev/null`" ]; then
+      installdir="`realpath \"$input\"`"
+    else
+      installdir="$input"
+    fi
   fi
   echo "Installing to $installdir"
-  echo -n "Continue? [Y/n/q] "
+  if [ ! -e "$installdir" ]; then
+    echo "NOTE: $installdir will be created"
+  elif [ -e "$installdir/$STATEEXE" ]; then
+    echo "WARNING: overwriting previous installation"
+  fi
+  echo -n "Continue? [y/N/q] "
   read response
   case "$response" in
     [Qq])
@@ -73,6 +119,9 @@ while "true"; do
       ;;
     [Yy])
       # Install.
+      if [ ! -e "$installdir" ]; then
+        mkdir -p "$installdir" || continue
+      fi
       echo "Installing to $installdir..."
       mv $stateexe "$installdir/$STATEEXE"
       if [ $? -eq 0 ]; then
@@ -85,10 +134,30 @@ while "true"; do
   esac
 done
 
-# Update user's profile and the current environment.
-echo "Updating environment..."
+# Check if the installation is in $PATH, if not, update user's profile if
+# permitted to.
+if [ "$installdir" = "`echo \"$PATH\" | grep -Fo \"$installdir\"`" ]; then
+  echo "Installation complete."
+  echo "You may now start using the 'state' program."
+  exit 0
+fi
 profile="`echo $HOME`/.profile"
-pathenv="export PATH=\"\$PATH:$installdir\" #ActiveState"
+if [ ! -w "$profile" ]; then
+  echo "Installation complete."
+  echo "Please manually add $installdir to your \$PATH in order to start"
+  echo "using the 'state' program."
+  exit 1
+fi
+echo -n "Allow \$PATH to be appended to in your $profile? [Y/n]"
+read response
+if [ "$response" == "Y" -o "$response" == "y" ]; then
+  echo "Installation complete."
+  echo "Please manually add $installdir to your \$PATH in order to start"
+  echo "using the 'state' program."
+  exit 1
+fi
+echo "Updating environment..."
+pathenv="export PATH=\"\$PATH:$installdir\" #$STATEID"
 if [ -z "`grep -no \"\#$STATEID\" \"$profile\"`" ]; then
   echo "Adding to \$PATH in $profile"
   echo "\n$pathenv" >> "$profile"
@@ -97,6 +166,6 @@ else
   sed -i -e "s|^export PATH=[^\#]\+\#$STATEID|$pathenv|;" "$profile"
 fi
 
-echo "Done."
+echo "Installation complete."
 echo "Please either run 'source ~/.profile' or start a new login shell in "
-echo "order to complete the installation."
+echo "order to start using the 'state' program."
